NEW require member password change on next login admin option
Add an option to the member adminstration form to require that user to change their password the next time they log in.
This is handled via existing functionality within the SilverStripe core by setting the password expirty date to 'now'.

In this manner an admin can set up a user and not know their password in the case that the user doesn't reset it, by simply forcing the reset.