NEW allow unsetting change password on next log in
Admins could require a user to reset their password on next log in, however once set there was no way to tell it was set as such, nor reverse that decision. To this end it is now possible to see the current state of whether or not a user will be required to change their password on logging in through the checkbox added previously to require this action.

It is now also possible to uncheck that checkbox and save the member's record to set the password expiry to NULL.

Docs were added to the readme recording this information in a more accessible manner.