FIX point sudomode help link to userhelp not google
It's not very useful to have a hardcoded help link that points to Google!

I found a reference to 'http://google.com' in the test `client/src/containers/SudoMode/tests/SudoMode-test.js` as well but decided to leave that there given it seems to just check for the presence of a link.